JusticeAssist-Deploy/backend/justiceassist/requirements.txt
# Core Flask
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5

# Database
sqlalchemy==2.0.20

# Security & Auth
passlib==1.7.4
werkzeug==2.3.7

# HTTP & Network
requests==2.31.0
ipwhois==1.2.0
dnspython==2.0.0
python-whois==0.8.0

# OCR & Image Processing
Pillow==10.0.0
pytesseract==0.3.10
piexif==1.1.3

# PDF Processing
PyMuPDF==1.23.5
PyPDF2==3.0.1
python-docx==0.8.11

# AI APIs
google-generativeai==0.3.2
google-genai==0.2.2
google-cloud-vision==3.1.2
openai>=1.55.0

# Task Queue & Background Jobs
celery==5.3.6
redis==5.0.1
rq==1.15.1

# Utilities
python-dotenv==1.0.0
phonenumbers==8.13.22
tldextract==5.1.2
deep-translator==1.11.4
tavily-python==0.3.0
langdetect==1.0.9

# NLP (Warning: Large packages - may slow Docker builds)
transformers==4.35.0
torch==2.1.0

# Testing & Development
pytest==7.4.2
black==23.7.0
flake8==6.1.0

gunicorn==21.2.0